feat(RS-01-34): implement checking access

+++ src/modules/posts/modules/versions/controllers/versions.controller.ts
@@ -1,24 +1,49 @@
 import { QueryParamsDto } from '@/common/dto/query-params.dto';
 import { PaginatedResponse } from '@/common/interfaces/pagination.interface';
-import { Author, PostVersion } from '@/generated/prisma';
+import { PostVersion } from '@/generated/prisma';
 import { PostModel } from '@/generated/prisma/models';
 import { JwtAccessGuard } from '@/modules/auth/guards/jwt-acess.guard';
+import { POST_ACCESS_LEVEL } from '@/modules/posts/constants/post';
 import { PostVersionDiff } from '@/modules/posts/modules/versions/dto/post-version-diff';
+import { PostsAccessService } from '@/modules/posts/services/posts-access.service';
+import { FullUser } from '@/modules/users/types/user.type';
 import { CurrentUser } from '@/shared/decorators/current-user.decorator';
-import { Controller, Delete, Get, Param, ParseIntPipe, Post, Query, UseGuards } from '@nestjs/common';
+import {
+  Controller,
+  Delete,
+  ForbiddenException,
+  Get,
+  Param,
+  ParseIntPipe,
+  Post,
+  Query,
+  UseGuards,
+} from '@nestjs/common';
 
 import { VersionsService } from '../services/versions.service';
 
 @UseGuards(JwtAccessGuard)
 @Controller('posts/:postId/versions')
 export class PostVersionsController {
-  constructor(private readonly versionsService: VersionsService) {}
+  constructor(
+    private postsAccessService: PostsAccessService,
+    private readonly versionsService: VersionsService,
+  ) {}
 
   @Get()
   public async getVersionsByPost(
     @Param('postId', ParseIntPipe) postId: number,
     @Query() query: QueryParamsDto,
+    @CurrentUser() user: FullUser | null,
   ): Promise<PaginatedResponse<PostVersion>> {
+    const hasAccess = await this.postsAccessService.hasAccess(
+      user,
+      { id: postId },
+      POST_ACCESS_LEVEL.AUTHOR_OR_MODERATOR,
+    );
+    if (!hasAccess) {
+      throw new ForbiddenException();
+    }
     return this.versionsService.getVersionsByPost(postId, query);
   }
 
@@ -26,16 +51,33 @@ export class PostVersionsController {
   public async revertToVersion(
     @Param('postId', ParseIntPipe) postId: number,
     @Param('version', ParseIntPipe) version: number,
-    @CurrentUser('author') author: Author,
+    @CurrentUser() user: FullUser | null,
   ): Promise<PostModel> {
-    return this.versionsService.revertToVersion(postId, version, author);
+    const hasAccess = await this.postsAccessService.hasAccess(
+      user,
+      { id: postId },
+      POST_ACCESS_LEVEL.MAIN_AUTHOR_OR_MODERATOR,
+    );
+    if (!hasAccess) {
+      throw new ForbiddenException();
+    }
+    return this.versionsService.revertToVersion(postId, version, user?.author);
   }
 
   @Delete(':version')
   public async deletePostVersion(
     @Param('postId', ParseIntPipe) postId: number,
     @Param('version', ParseIntPipe) version: number,
+    @CurrentUser() user: FullUser | null,
   ): Promise<PostVersion> {
+    const hasAccess = await this.postsAccessService.hasAccess(
+      user,
+      { id: postId },
+      POST_ACCESS_LEVEL.MAIN_AUTHOR_OR_MODERATOR,
+    );
+    if (!hasAccess) {
+      throw new ForbiddenException();
+    }
     return this.versionsService.deletePostVersion(postId, version);
   }
 
@@ -44,7 +86,16 @@ export class PostVersionsController {
     @Param('postId', ParseIntPipe) postId: number,
     @Query('from', ParseIntPipe) from: number,
     @Query('to', ParseIntPipe) to: number,
+    @CurrentUser() user: FullUser | null,
   ): Promise<PostVersionDiff> {
+    const hasAccess = await this.postsAccessService.hasAccess(
+      user,
+      { id: postId },
+      POST_ACCESS_LEVEL.AUTHOR_OR_MODERATOR,
+    );
+    if (!hasAccess) {
+      throw new ForbiddenException();
+    }
     return this.versionsService.comparePostVersions(postId, from, to);
   }
 }
